[Fix] #520 - toPresentation() 함수 viewModel에서 PresentationModel로 이동
