+++ SOPT-iOS/Projects/Features/SoptlogFeature/Sources/SoptlogScene/ViewModel/SoptlogViewModel.swift
@@ -107,46 +107,3 @@ extension SoptlogViewModel {
         return output
     }
 }
-
-extension SoptlogModel {
-    func toPresentation() -> SoptlogPresentationModel {
-        var appService: [SoptlogPresentationModel.AppService] = []
-        appService.append(SoptlogPresentationModel.AppService(
-            serviceName: I18N.Soptlog.soptlevel,
-            serviceImageURL: self.icons[0],
-            serviceValue: self.soptLevel))
-        appService.append(SoptlogPresentationModel.AppService(
-            serviceName: I18N.Soptlog.poke,
-            serviceImageURL: self.icons[1],
-            serviceValue: self.pokeCount))
-        
-        if self.isActive {
-            appService.append(SoptlogPresentationModel.AppService(
-                serviceName: I18N.Soptlog.soptamp,
-                serviceImageURL: self.icons[2],
-                serviceValue: self.soptampRank))
-        } else {
-            appService.append(SoptlogPresentationModel.AppService(
-                serviceName: I18N.Soptlog.withSopt,
-                serviceImageURL: self.icons[2],
-                serviceValue: self.during))
-        }
-        
-        
-        return SoptlogPresentationModel(
-            profile: SoptlogPresentationModel.Profile(
-                userName: self.userName,
-                profileImage: self.profileImage,
-                part: self.part
-            ),
-            introduce: SoptlogPresentationModel.Introduce(
-                profileMessage: self.profileMessage
-            ),
-            appService: appService,
-            alarm: SoptlogPresentationModel.Alarm(
-                isFortuneChecked: self.isFortuneChecked,
-                todayFortuneText: self.todayFortuneText
-            )
-        )
-    }
-}
